# docs/swagger.yml
openapi: 3.1.0
info:
  title: Vectis Backend API
  version: 0.1.0
  summary: Tenant-aware API for the Vectis platform backend.
  description: >
    OpenAPI specification for the current Vectis backend. The platform currently
    exposes a mix of seeded overview/dashboard endpoints and persisted tenant-scoped
    CRUD endpoints for premises and cameras. Local development uses PostgreSQL,
    and production is intended to use PostgreSQL via AWS RDS.
servers:
  - url: http://localhost:3000
    description: Local backend server
...
paths:
...
  /api/v1/premises:
    get:
      tags:
        - Premises
      summary: List premises for the current tenant
      operationId: listPremises
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: Tenant-scoped premises.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    post:
      tags:
        - Premises
      summary: Create a premises
      operationId: createPremises
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePremisesRequest"
      responses:
        "201":
          description: Premises created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/premises/{premisesId}:
    get:
      tags:
        - Premises
      summary: Get a premises by id
      operationId: getPremises
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PremisesId"
      responses:
        "200":
          description: Premises found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    patch:
      tags:
        - Premises
      summary: Update a premises
      operationId: updatePremises
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/PremisesId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePremisesRequest"
      responses:
        "200":
          description: Premises updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    delete:
      tags:
        - Premises
      summary: Delete a premises
      operationId: deletePremises
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/PremisesId"
      responses:
        "204":
          description: Premises deleted.
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/premises/{premisesId}/cameras:
    get:
      tags:
        - Cameras
      summary: List cameras for a premises
      operationId: listPremisesCameras
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PremisesId"
      responses:
        "200":
          description: Cameras for the premises.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesCameraListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    post:
      tags:
        - Cameras
      summary: Create a camera under a premises
      operationId: createPremisesCamera
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/PremisesId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCameraRequest"
      responses:
        "201":
          description: Camera created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesCameraResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/premises/{premisesId}/cameras/{cameraId}:
    get:
      tags:
        - Cameras
      summary: Get a camera by id
      operationId: getPremisesCamera
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PremisesId"
        - $ref: "#/components/parameters/CameraId"
      responses:
        "200":
          description: Camera found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesCameraResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    patch:
      tags:
        - Cameras
      summary: Update a camera
      operationId: updatePremisesCamera
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/PremisesId"
        - $ref: "#/components/parameters/CameraId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCameraRequest"
      responses:
        "200":
          description: Camera updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremisesCameraResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    delete:
      tags:
        - Cameras
      summary: Delete a camera
      operationId: deletePremisesCamera
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/PremisesId"
        - $ref: "#/components/parameters/CameraId"
      responses:
        "204":
          description: Camera deleted.
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/agents:
    get:
      tags:
        - Agents
      summary: List agents for the current tenant
      operationId: listAgents
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: Tenant-scoped agents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
    post:
      tags:
        - Agents
      summary: Create an agent
      operationId: createAgent
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get an agent by id
      operationId: getAgent
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    patch:
      tags:
        - Agents
      summary: Update an agent
      operationId: updateAgent
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: Agent updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
    delete:
      tags:
        - Agents
      summary: Delete an agent
      operationId: deleteAgent
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/UserHeader"
        - $ref: "#/components/parameters/AgentId"
      responses:
        "204":
          description: Agent deleted.
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
...
  /api/v1/agents/{agentId}/frames:
    post:
      tags:
        - Ingest
      summary: Submit multiple filtered frames from an agent
      operationId: ingestAgentFrames
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestFramesRequest"
      responses:
        "202":
          description: Frames accepted and persisted to storage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraFrameListResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
components:
  parameters:
    TenantHeader:
      name: x-tenant-id
      in: header
      required: true
      description: Tenant account identifier used to scope protected resources.
      schema:
        type: string
    UserHeader:
      name: x-user-id
      in: header
      required: true
      description: Acting user identifier for write operations.
      schema:
        type: string
    PremisesId:
      name: premisesId
      in: path
      required: true
      schema:
        type: string
    CameraId:
      name: cameraId
      in: path
      required: true
      schema:
        type: string
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
  responses:
    UnauthorizedResponse:
      description: Request is missing tenant or user context.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorized
            message: Tenant context is required for this resource.
    BadRequestResponse:
      description: Request payload failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundResponse:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PremisesListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Premises"
    PremisesResponse:
      type: object
      required:
        - premises
      properties:
        premises:
          $ref: "#/components/schemas/Premises"
    PremisesCameraListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PremisesCamera"
    PremisesCameraResponse:
      type: object
      required:
        - camera
      properties:
        camera:
          $ref: "#/components/schemas/PremisesCamera"
    AgentListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
    AgentResponse:
      type: object
      required:
        - agent
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
...
    CameraFrameListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CameraFrame"
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    PremisesType:
      type: string
      enum:
        - house
        - apartment
        - office
        - factory
        - warehouse
        - retail
        - other
    CameraStatus:
      type: string
      enum:
        - online
        - degraded
        - offline
        - maintenance
    AgentStatus:
      type: string
      enum:
        - online
        - offline
        - maintenance
    Premises:
      type: object
      required:
        - id
        - tenantId
        - name
        - type
        - addressLine1
        - city
        - countryCode
        - createdByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/PremisesType"
        addressLine1:
          type: string
        addressLine2:
          type:
            - string
            - "null"
        city:
          type: string
        state:
          type:
            - string
            - "null"
        postalCode:
          type:
            - string
            - "null"
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        notes:
          type:
            - string
            - "null"
        createdByUserId:
          type: string
        updatedByUserId:
          type:
            - string
            - "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PremisesCamera:
      type: object
      required:
        - id
        - tenantId
        - premisesId
        - name
        - streamUrl
        - status
        - createdByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        premisesId:
          type: string
        name:
          type: string
        streamUrl:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/CameraStatus"
        model:
          type:
            - string
            - "null"
        serialNumber:
          type:
            - string
            - "null"
        locationDescription:
          type:
            - string
            - "null"
        createdByUserId:
          type: string
        updatedByUserId:
          type:
            - string
            - "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Agent:
      type: object
      required:
        - id
        - tenantId
        - premisesId
        - name
        - status
        - createdByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        premisesId:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/AgentStatus"
        softwareVersion:
          type:
            - string
            - "null"
        locationDescription:
          type:
            - string
            - "null"
        hostName:
          type:
            - string
            - "null"
        createdByUserId:
          type: string
        updatedByUserId:
          type:
            - string
            - "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CameraFrame:
      type: object
      required:
        - id
        - tenantId
        - agentId
        - cameraId
        - storagePath
        - contentType
        - byteSize
        - capturedAt
        - receivedAt
      properties:
        id:
          type: string
        tenantId:
          type: string
        agentId:
          type: string
        cameraId:
          type: string
        storagePath:
          type: string
        contentType:
          type: string
        byteSize:
          type: integer
        capturedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
    CreatePremisesRequest:
      type: object
      required:
        - name
        - type
        - addressLine1
        - city
        - countryCode
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: "#/components/schemas/PremisesType"
        addressLine1:
          type: string
          minLength: 1
        addressLine2:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        postalCode:
          type: string
          minLength: 1
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        notes:
          type: string
          minLength: 1
    UpdatePremisesRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: "#/components/schemas/PremisesType"
        addressLine1:
          type: string
          minLength: 1
        addressLine2:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        postalCode:
          type: string
          minLength: 1
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        notes:
          type: string
          minLength: 1
    CreateCameraRequest:
      type: object
      required:
        - name
        - streamUrl
      properties:
        name:
          type: string
          minLength: 1
        streamUrl:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/CameraStatus"
        model:
          type: string
          minLength: 1
        serialNumber:
          type: string
          minLength: 1
        locationDescription:
          type: string
          minLength: 1
    UpdateCameraRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        streamUrl:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/CameraStatus"
        model:
          type: string
          minLength: 1
        serialNumber:
          type: string
          minLength: 1
        locationDescription:
          type: string
          minLength: 1
    CreateAgentRequest:
      type: object
      required:
        - premisesId
        - name
      properties:
        premisesId:
          type: string
        name:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/AgentStatus"
        softwareVersion:
          type: string
          minLength: 1
        locationDescription:
          type: string
          minLength: 1
        hostName:
          type: string
          minLength: 1
    UpdateAgentRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        status:
          $ref: "#/components/schemas/AgentStatus"
        softwareVersion:
          type: string
          minLength: 1
        locationDescription:
          type: string
          minLength: 1
        hostName:
          type: string
          minLength: 1
...
    IngestFramesRequest:
      type: object
      required:
        - frames
      properties:
        frames:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - cameraId
              - timestamp
              - contentType
              - dataBase64
            properties:
              cameraId:
                type: string
              timestamp:
                type: string
                format: date-time
              contentType:
                type: string
                enum:
                  - image/jpeg
                  - image/png
              dataBase64:
                type: string
                minLength: 1